backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.35
alembic==1.13.2
pydantic==2.9.2
python-dotenv==1.0.1
apscheduler==3.10.4
selenium==4.24.0
webdriver-manager==4.0.2
email-validator==2.2.0
python-telegram-bot==21.6
requests==2.32.3
httpx==0.27.2
